New Development: Yes/No             JIRA Issue: CAS-6510         Ready for Test: Yes      Interface Changes: Yes What Interface Changed: Add ImageFactory::imageFromArray(), remove ImageAnalysis::imagefromarray()          Test Programs: L   Put in Release Notes: Yes/No              Module(s): Module Names change impacts.            Description: as above, imageFromArray supports complex valued images
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@29659 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/imageanalysis/ImageAnalysis/PixelValueManipulator.tcc
@@ -135,9 +135,9 @@ template<class T> Record PixelValueManipulator<T>::getProfile(
 		 Vector<Double> inputRef = this->_getImage()->coordinates().referenceValue();
 		 inputRef[axis] = outputRef[axis];
 		 Vector<Double> inputPixel = this->_getImage()->coordinates().toPixel(inputRef);
-		 Int count = floor(inputPixel[axis] + 0.5);
+		 Double count = floor(inputPixel[axis] + 0.5);
 		 uInt length = values.shape()[0];
-		 Vector<Int> coords = indgen(length, count, 1);
+		 Vector<Double> coords = indgen(length, count, 1.0);
 		 ret.define("coords", coords);
 		 ret.define("xUnit", "pixel");
 	 }
